Added the implementation of adding the Lipa na mpesa data.

+++ database/migrations/2020_09_13_043329_create_lipa_na_mpesas_table.php
@@ -17,7 +17,7 @@ class CreateLipaNaMpesasTable extends Migration
             $table->id();
             $table->longText('MerchantRequestID')->nullable();
             $table->longText('CheckoutRequestID')->nullable();
-            $table->bigInteger('Amount')->nullable();
+            $table->float('Amount')->nullable();
             $table->longText('MpesaReceiptNumber')->nullable();
             $table->longText('TransactionDate')->nullable();
             $table->longText('PhoneNumber')->nullable();
